More BASH_REMATCH fixes
We need to save the args as they were at the time of the BASH_REMATCH
command, not as it might be subsequent times it needs to be checked.

It is possible the debugger sets BASH_REMATCH on its own when it
should be empty. So check for that as well.